AMB-1325 add checks for decode64
