+++ proxies/live/apiproxy/policies/AssignMessage.400InvalidBase64Encoding.xml
@@ -1,0 +1,32 @@
+<AssignMessage async="false" continueOnError="false" enabled="true" name="AssignMessage.400InvalidBase64Encoding">
+  <Set>
+    <StatusCode>400</StatusCode>
+    <Payload contentType="application/json">
+        {
+            "resourceType": "OperationOutcome",
+            "meta": {
+                "profile": [
+                "https://fhir.hl7.org.uk/StructureDefinition/UKCore-OperationOutcome"
+                ]
+            },
+            "issue": [
+                {
+                "severity": "error",
+                "code": "400",
+                "details": {
+                    "coding": [
+                    {
+                        "system": "http://hl7.org/fhir/ValueSet/operation-outcome",
+                        "code": "PROXY_BAD_REQUEST"
+                    }
+                    ]
+                },
+                "diagnostics": "The base64 decoding of the NHSD-Target-Identifier header failed."
+                }
+            ]
+        }
+    </Payload>
+  </Set>
+  <IgnoreUnresolvedVariables>true</IgnoreUnresolvedVariables>
+  <AssignTo createNew="false" transport="http" type="response" />
+</AssignMessage>
