+++ tests/test_appointment.py
@@ -18,7 +18,6 @@ class TestAppointment:
     @pytest.mark.appointment
     @pytest.mark.integration
     @pytest.mark.sandbox
-    @pytest.mark.debug
     def test_get_appointments(self, get_token_client_credentials):
         """
            test for /appointment..  to get all appointments for the patient passed as parameter on the request - /Appointment?patient:identifier=12312
